HBASE-5945 Reduce buffer copies in IPC server response path
git-svn-id: https://svn.apache.org/repos/asf/hbase/trunk@1543458 13f79535-47bb-0310-9956-ffa450edef68

+++ hbase-hadoop2-compat/src/main/java/org/apache/hadoop/hbase/ipc/MetricsHBaseServerSourceImpl.java
@@ -89,7 +89,7 @@ public class MetricsHBaseServerSourceImpl extends BaseSourceImpl
   }
 
   @Override
-  public void sentBytes(int count) {
+  public void sentBytes(long count) {
     this.sentBytes.incr(count);
   }
