+++ hbase-server/src/main/java/org/apache/hadoop/hbase/ipc/MetricsHBaseServer.java
@@ -47,7 +47,7 @@ public class MetricsHBaseServer {
     source.authenticationSuccess();
   }
 
-  void sentBytes(int count) {
+  void sentBytes(long count) {
     source.sentBytes(count);
   }
